refactor: Refactor delete article
